impl handler layer and auth middlewares

+++ internal/model/house.go
@@ -4,9 +4,9 @@ import "time"
 
 // Дом
 type House struct {
-	ID                 int       `json:"id"`
+	ID                 int64     `json:"id"`
 	Address            string    `json:"address"`
-	YearOfConstruction int       `json:"year_of_construction"`
+	YearOfConstruction int64     `json:"year_of_construction"`
 	Developer          *string   `json:"developer,omitempty"`
 	CreatedAt          time.Time `json:"created_at"`
 	UpdatedAt          time.Time `json:"updated_at"`
